Add metagame setting to restrict party ally visibility
