+++ static/lang/en.json
@@ -2849,6 +2849,10 @@
                     "Hint": "If enabled, players will see the degree of success of checks made against NPCs and other non-player-owned sources.",
                     "Name": "Show results on attacks and saves"
                 },
+                "ShowPartyStats": {
+                    "Hint": "If enabled, players will see stats for their fellow party members in the party actor sheet",
+                    "Name": "Show Party Member Stats"
+                },
                 "TokenSetsNameVisibility": {
                     "Hint": "If enabled, then for any NPC token whose nameplate isn't visible to players, its name will also be hidden from them in the encounter tracker and chat messages.",
                     "Name": "Tokens determine NPC name visibility"
