+++ static/templates/actors/party/regions/inventory-members.hbs
@@ -1,34 +1,40 @@
 {{#if members}}
     <ol class="inventory-members">
-        <li class="box summary">
-            <header>{{localize "PF2E.Actor.Party.Total"}}</header>
-            <div class="summary-data">
-                <div>
-                    <label><i class="fa-solid fa-coins fa-fw"></i> {{localize "PF2E.Actor.Party.Coin"}}</label>
-                    <span class="value">{{numberFormat inventorySummary.totalCoins decimals=2}} gp</span>
+        {{#unless restricted}}
+            <li class="box summary">
+                <header>{{localize "PF2E.Actor.Party.Total"}}</header>
+                <div class="summary-data">
+                    <div>
+                        <label><i class="fa-solid fa-coins fa-fw"></i> {{localize "PF2E.Actor.Party.Coin"}}</label>
+                        <span class="value">{{numberFormat inventorySummary.totalCoins decimals=2}} gp</span>
+                    </div>
+                    <div>
+                        <label><i class="fa-solid fa-scale-unbalanced fa-fw"></i> {{localize "PF2E.Actor.Party.Wealth"}}</label>
+                        <span class="value">{{numberFormat inventorySummary.totalWealth decimals=2}} gp</span>
+                    </div>
                 </div>
-                <div>
-                    <label><i class="fa-solid fa-scale-unbalanced fa-fw"></i> {{localize "PF2E.Actor.Party.Wealth"}}</label>
-                    <span class="value">{{numberFormat inventorySummary.totalWealth decimals=2}} gp</span>
-                </div>
-            </div>
-            <footer>
-                <i class="fa-solid fa-weight-hanging"></i> {{localize "PF2E.BulkShortLabel"}} {{inventorySummary.totalBulk}}
-            </footer>
-        </li>
+                <footer>
+                    <i class="fa-solid fa-weight-hanging"></i> {{localize "PF2E.BulkShortLabel"}} {{inventorySummary.totalBulk}}
+                </footer>
+            </li>
+        {{/unless}}
         {{#each members as |member|}}
             <li class="box">
                 <div class="actor-link content" data-actor-uuid="{{member.actor.uuid}}" data-action="open-sheet" data-tab="inventory">
                     <img src="{{member.actor.img}}" />
                     <span class="name">{{member.actor.name}}</span>
-                    <span class="value">
-                        <i class="fa-solid fa-scale-unbalanced"></i>
-                        {{numberFormat member.actor.inventory.totalWealth.goldValue decimals=2}} gp
-                    </span>
+                    {{#unless member.restricted}}
+                        <span class="value">
+                            <i class="fa-solid fa-scale-unbalanced"></i>
+                            {{numberFormat member.actor.inventory.totalWealth.goldValue decimals=2}} gp
+                        </span>
+                    {{/unless}}
                 </div>
-                <footer>
-                    <i class="fa-solid fa-weight-hanging"></i> {{localize "PF2E.BulkShortLabel"}} {{member.actor.inventory.bulk.value}}
-                </footer>
+                {{#unless member.restricted}}
+                    <footer>
+                        <i class="fa-solid fa-weight-hanging"></i> {{localize "PF2E.BulkShortLabel"}} {{member.actor.inventory.bulk.value}}
+                    </footer>
+                {{/unless}}
             </li>
         {{/each}}
     </ol>
