+++ static/templates/actors/party/sheet.hbs
@@ -14,7 +14,9 @@
     </header>
 
     <nav class="sub-nav">
+        {{#unless restricted}}
         <a data-tab="overview">{{localize "PF2E.Actor.Party.Tabs.Overview"}}</a>
+        {{/unless}}
         <a data-tab="inventory">{{localize "PF2E.Actor.Party.Tabs.Inventory"}}</a>
         {{#if orphaned}}
             <a data-tab="orphaned">{{localize "PF2E.Actor.Party.Tabs.Orphaned"}}</a>
@@ -22,8 +24,9 @@
     </nav>
 
     <section class="container">
-        <div class="tab" data-tab="overview" data-region="overview">
-        </div>
+        {{#unless restricted}}
+            <div class="tab" data-tab="overview" data-region="overview"></div>
+        {{/unless}}
 
         <div class="tab" data-tab="inventory">
             <aside class="sidebar">
